configurable onDrop
